Refatora configurações e remove classe User
Refatora configurações de entidades com a introdução de
`BaseEntityConfiguration` e `MoneyConfigurationExtensions` para
padronizar e reutilizar código. Remove a classe `User` e a pasta
`Interfaces` do projeto. Altera o tipo de `Amount` em `Money` de
`double` para `decimal` para maior precisão em cálculos financeiros.
Atualiza namespaces e configurações de tabelas para entidades como
`Expense`, `ExpenseCategory` e `ExpenseInstallment`. Adiciona
configuração para `FinancialTransaction`.

+++ Backend/Costium.Domain/ValueObjects/Money.cs
@@ -5,15 +5,15 @@ namespace Costium.Domain.ValueObjects;
 
 public sealed class Money : ValueObject
 {
-    public double Amount { get; private set; }
+    public decimal Amount { get; private set; }
     public Currency Currency { get; private set; }
 
-    private Money(double amount, Currency currency) 
+    private Money(decimal amount, Currency currency) 
     {
         Amount = amount;
         Currency = currency;
     }
-    public static Money Create(double amount, Currency currency)
+    public static Money Create(decimal amount, Currency currency)
     {
         if (amount < 0)
             throw new DomainException("Valor não pode ser negativo.");
